// types.ts
export enum AiOperation {
  SUMMARIZE = 'Summarize',
  REWRITE_FORMAL = 'Rewrite (Formal)',
  REWRITE_CASUAL = 'Rewrite (Casual)',
  EXPAND = 'Expand',
  CHECK_GRAMMAR_POLISH = 'Check Grammar & Polish',
  SIMPLIFY = 'Simplify Language',
  GENERATE_IDEAS = 'Generate Ideas (from topic)',
  TRANSLATE = 'Translate', // New generic translate operation
}